Introduced Posterior Model Builder, which replaces the PartiallyFixedModelBuilder.
We have refactored the PartiallyFixedModelBuilder such that the code closely corresponds to the theory and notation used in our Media paper "Posterior Shape Models". To make the connection more clear, we have saved it under the new name PosteriorModelBuilder. The PartiallyFixedModelBuilder is deprecated and will in future versions be removed.